requirements.txt
flask==2.0.1
flask-cors==3.0.10
pyyaml==6.0
selenium==4.10.0
webdriver-manager==3.8.6
